Add: Cell Merges
This adds cell merges that can be horizontal, vertical, or both. SlickGrid doesn't normally support vertical merges, so it's hacked in using some CSS. The risk this exposes is that when we go to add selection support, we'll need to make sure the selection model knows about these vertical merges (otherwise we'd get a weird selection that can cut through merges). We'd also want to make sure that it still works when copying.

Closes #1